main system app component is now pluggable
