+++ .eslintrc
@@ -3,6 +3,7 @@
 	"plugins": [ "flowtype" ],
 	"parser": "babel-eslint",
 	"rules": {
-		"react/react-in-jsx-scope": 0
+		"react/react-in-jsx-scope": 0, 
+		"prefer-arrow-callback": 0
 	}
 }
